# openapi.yaml
openapi: 3.0.1
info:
  title: NewsDataHub API
  description: |
    The NewsDataHub API provides access to news articles and sources with comprehensive filtering options.

    ## Authentication
    All endpoints require authentication using an API key in the `X-API-Key` header.

    **Get your API key:**
    Sign up at [newsdatahub.com](https://newsdatahub.com) to get instant access. Choose from:
    - **Free tier** with limited features (no credit card required)
    - **Full-featured 7-day free trial** (credit card required)

    ## API Quotas and Rate Limits
    API usage limits vary by subscription tier:
    - **Free**: 50 requests/day
    - **Free Trial**: 50 requests/day for 7 days (during trial period)
    - **Starter**: 300 requests/day (9,000 calls/month)
    - **Basic**: 200 requests/minute (20,000 calls/month)
    - **Pro**: 200 requests/minute (30,000 calls/month)

    ## Pagination
    All endpoints use cursor-based pagination for efficient handling of large result sets.

  version: 1.0.0
  contact:
    name: NewsDataHub Support
    email: support@newsdatahub.com
    url: https://newsdatahub.com
  license:
    name: Proprietary
    url: https://newsdatahub.com/terms-of-service
servers:
  - url: https://api.newsdatahub.com
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /v1/news/{article_id}/related:
    get:
      summary: Get related articles
      description: |
        Find articles similar to a given article.

        **Similarity matching**: Compares title, description, topics, and keywords to find related articles.

        **Recency filter**: Only returns articles published within the last 60 days.

        **Access Requirements:**
        - Requires a paid plan or active trial
        - Not available on the free tier

        **Content Access Levels:**
        - **Starter**: Full content, no keywords/topics/sentiment
        - **Professional/Business/Enterprise and legacy Basic/Pro**: Full content with keywords, topics, and sentiment analysis

      operationId: getRelatedArticles
      tags:
        - News
      parameters:
        - name: article_id
          in: path
          description: The ID (UUID) of the article to find similar articles for
          required: true
          schema:
            type: string
            format: uuid
            example: "fa90d4bd-f901-4ea6-8b4d-e6f7a451d311"
        - name: per_page
          in: query
          description: |
            Number of related articles to return (tier-dependent limits):
            - Free: max 10
            - Starter: max 30
            - Basic: max 50
            - Pro: max 100
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 5
            example: 5
        - name: fields
          in: query
          description: "Comma-separated list of fields to return (e.g., title,pub_date,source_title). Returns all fields if not specified. 'id' is always included."
          required: false
          schema:
            type: string
            example: "title,pub_date,source_title"
      responses:
        '200':
          description: Successful response with related articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  related_to:
                    type: object
                    description: Information about the source article that was queried
                    properties:
                      id:
                        type: string
                        description: The article ID (UUID)
                        example: "fa90d4bd-f901-4ea6-8b4d-e6f7a451d311"
                      title:
                        type: string
                        description: The article title
                        example: "Breaking Technology News"
                      source_title:
                        type: string
                        description: The news source name
                        example: "TechCrunch"
                      article_link:
                        type: string
                        description: Link to the original article
                        example: "https://techcrunch.com/article"
                      pub_date:
                        type: string
                        description: Publication date in ISO 8601 format
                        example: "2025-01-15T10:00:00Z"
                  count:
                    type: integer
                    description: Number of related articles returned
                    example: 5
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NewsArticle'
              example:
                related_to:
                  id: "fa90d4bd-f901-4ea6-8b4d-e6f7a451d311"
                  title: "Breaking Technology News"
                  source_title: "TechCrunch"
                  article_link: "https://techcrunch.com/article"
                  pub_date: "2025-01-15T10:00:00Z"
                count: 3
                data:
                  - id: "c6d1fc78-2ff3-43e6-9889-2da0ea831262"
                    title: "Similar Technology Story"
                    source_title: "Boston.com"
                    pub_date: "2025-01-15T09:00:00Z"
                  - id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    title: "Another Related Article"
                    source_title: "WHDH 7News"
                    pub_date: "2025-01-14T18:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    NewsArticle:
      type: object
      required:
        - id
        - title
        - source_title
        - source_link
        - article_link
        - description
        - pub_date
        - language
      properties:
        id:
          type: string
          description: Unique article identifier
          example: "c6d1fc78-2ff3-43e6-9889-2da0ea831262"
        title:
          type: string
          description: Article headline
          example: "Major Technology Breakthrough Announced"
        source_title:
          type: string
          description: Name of the news source
          example: "TechNews Daily"
        source_link:
          type: string
          format: uri
          description: URL of the news source
          example: "https://technews.com"
        article_link:
          type: string
          format: uri
          description: Direct URL to the article
          example: "https://technews.com/articles/breakthrough"
        keywords:
          type: array
          items:
            type: string
          description: Keywords associated with the article (Basic/Pro tiers only)
          example: ["technology", "innovation", "AI"]
        topics:
          type: array
          items:
            type: string
          description: Detected topics and categories (Basic/Pro tiers only)
          example: ["technology", "artificial intelligence"]
        description:
          type: string
          description: Article description or excerpt
          example: "Scientists announce significant breakthrough in AI technology..."
        pub_date:
          type: string
          format: date-time
          description: Article publication timestamp (ISO 8601)
          example: "2024-01-15T10:30:00Z"
        creator:
          type: string
          nullable: true
          description: Article author
          example: "John Smith"
        content:
          type: string
          description: Full article content (truncated for free tier)
          example: "Scientists at leading research institutions have announced..."
        media_url:
          type: string
          format: uri
          nullable: true
          description: URL to associated media
          example: "https://technews.com/images/breakthrough.jpg"
        media_type:
          type: string
          nullable: true
          description: Type of media associated with the article
          example: "image"
        language:
          type: string
          description: Language code (2-letter ISO)
          example: "en"
        sentiment:
          $ref: '#/components/schemas/Sentiment'
        source:
          $ref: '#/components/schemas/SourceMetadata'
    Sentiment:
      type: object
      description: Sentiment analysis scores (Basic/Pro tiers only)
      properties:
        pos:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Positive sentiment score
          example: 0.7
        neg:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Negative sentiment score
          example: 0.1
        neu:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Neutral sentiment score
          example: 0.2
    SourceMetadata:
      type: object
      properties:
        id:
          type: string
          description: Source identifier
          example: "technews-daily"
        country:
          type: string
          description: Source country code (2-letter ISO)
          example: "US"
        political_leaning:
          type: string
          description: Political leaning classification
          example: "center"
        reliability_score:
          type: number
          format: float
          minimum: 0
          maximum: 10
          description: Reliability score (0-10)
          example: 8.5
        type:
          type: string
          description: Source type
          example: "digital_native"
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error type or brief description
          example: "Bad Request"
        message:
          type: string
          description: Detailed error message describing what went wrong
          example: "Invalid country code format"
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Bad Request"
            message: "Invalid country code format"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            message: "API key is either invalid or expired/deactivated"
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Too Many Requests"
            message: "You have exceeded your monthly/daily quota or per-minute rate limit. Check the X-Quota-* and X-RateLimit-* headers for your current usage and limits, or upgrade to a higher quota tier."
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal Server Error"
            message: "An unexpected error occurred on the server. Please try again later."
...